replaced the form page with the content of the reworked one and removed the temp reworked one
