+++ frontend/app/pages/admin/productions/create.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts"></script>
 
 <template>
-  <AdminProductionsFormPageNew mode="create" />
+  <AdminProductionsFormPage mode="create" />
 </template>
 
 <style scoped></style>
